# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Used Fruit API",
    "version": "1.0.0-staging",
    "description": "Implementierungsvertrag. Noch nicht öffentlich freigegeben. IDs sind serverseitige UUIDs; Anzeigenummern separate unveränderliche Dezimalstrings. Benutzerzuordnung ausschließlich über IDs. Fehler enthalten keine privaten Daten. Web nutzt HttpOnly-Cookie und Origin-Prüfung; iOS widerrufbare Bearer-Tokens. Produktionsfreigabe erst nach Abnahme."
  },
  "servers": [
    {
      "url": "https://staging.usedfruit.de/api/v1",
      "description": "Aktive isolierte Abnahmeumgebung"
    },
    {
      "url": "https://usedfruit.de/api/v1",
      "description": "Erst nach Abnahme aktivieren"
    }
  ],
  "paths": {
    "/auth/apple/challenge": {
      "post": {
        "summary": "Einmalige Apple-Nonce, fünf Minuten gültig",
        "operationId": "post_auth_apple_challenge",
        "security": [],
        "responses": {
          "200": {
            "description": "Erfolgreich",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "challengeId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "nonceHash": {
                      "type": "string"
                    },
                    "expiresIn": {
                      "type": "integer"
                    }
                  },
                  "required": [
                    "challengeId",
                    "nonceHash",
                    "expiresIn"
                  ]
                },
                "example": {
                  "challengeId": "0847b184-a073-488a-ad76-b0cd6f7d9554",
                  "nonceHash": "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                  "expiresIn": 300
                }
              }
            }
          },
          "400": {
            "description": "Ungültige Eingabe",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Anmeldung erforderlich",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Keine Berechtigung",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Nicht gefunden oder gegenseitig gesperrt",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Versions- oder Idempotenzkonflikt",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Zu viele Anfragen",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Interner Fehler",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Fehler; keine privaten Daten.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": {
                    "code": "invalid_input",
                    "message": "Bitte prüfe deine Eingabe.",
                    "requestId": "0847b184-a073-488a-ad76-b0cd6f7d9554"
                  }
                }
              }
            }
          },
          "415": {
            "description": "Fehler; keine privaten Daten.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": {
                    "code": "invalid_input",
                    "message": "Bitte prüfe deine Eingabe.",
                    "requestId": "0847b184-a073-488a-ad76-b0cd6f7d9554"
                  }
                }
              }
            }
          },
          "503": {
            "description": "Fehler; keine privaten Daten.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": {
                    "code": "invalid_input",
                    "message": "Bitte prüfe deine Eingabe.",
                    "requestId": "0847b184-a073-488a-ad76-b0cd6f7d9554"
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
    "/auth/apple": {
      "post": {
        "summary": "Native Apple-Anmeldung: challengeId, identityToken, authorizationCode, optional name",
        "operationId": "post_auth_apple",
        "security": [],
        "responses": {
          "200": {
            "description": "Erfolgreich",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Tokens"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "user": {
                          "$ref": "#/components/schemas/Profile"
                        }
                      },
                      "required": [
                        "user"
                      ]
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Ungültige Eingabe",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Anmeldung erforderlich",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Keine Berechtigung",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Nicht gefunden oder gegenseitig gesperrt",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Versions- oder Idempotenzkonflikt",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Zu viele Anfragen",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Interner Fehler",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Fehler; keine privaten Daten.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": {
                    "code": "invalid_input",
                    "message": "Bitte prüfe deine Eingabe.",
                    "requestId": "0847b184-a073-488a-ad76-b0cd6f7d9554"
                  }
                }
              }
            }
          },
          "415": {
            "description": "Fehler; keine privaten Daten.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": {
                    "code": "invalid_input",
                    "message": "Bitte prüfe deine Eingabe.",
                    "requestId": "0847b184-a073-488a-ad76-b0cd6f7d9554"
                  }
                }
              }
            }
          },
          "503": {
            "description": "Fehler; keine privaten Daten.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": {
                    "code": "invalid_input",
                    "message": "Bitte prüfe deine Eingabe.",
                    "requestId": "0847b184-a073-488a-ad76-b0cd6f7d9554"
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "challengeId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "identityToken": {
                    "type": "string"
                  },
                  "authorizationCode": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string",
                    "maxLength": 40
                  }
                },
                "required": [
                  "challengeId",
                  "identityToken",
                  "authorizationCode"
                ]
              }
            }
          }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "summary": "Rotierendes Erneuerungstoken; Wiederverwendung widerruft die Tokenfamilie",
        "operationId": "post_auth_refresh",
        "security": [],
        "responses": {
          "200": {
            "description": "Erfolgreich",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tokens"
                }
              }
            }
          },
          "400": {
            "description": "Ungültige Eingabe",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Anmeldung erforderlich",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Keine Berechtigung",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Nicht gefunden oder gegenseitig gesperrt",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Versions- oder Idempotenzkonflikt",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Zu viele Anfragen",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Interner Fehler",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Fehler; keine privaten Daten.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": {
                    "code": "invalid_input",
                    "message": "Bitte prüfe deine Eingabe.",
                    "requestId": "0847b184-a073-488a-ad76-b0cd6f7d9554"
                  }
                }
              }
            }
          },
          "415": {
            "description": "Fehler; keine privaten Daten.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": {
                    "code": "invalid_input",
                    "message": "Bitte prüfe deine Eingabe.",
                    "requestId": "0847b184-a073-488a-ad76-b0cd6f7d9554"
                  }
                }
              }
            }
          },
          "503": {
            "description": "Fehler; keine privaten Daten.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": {
                    "code": "invalid_input",
                    "message": "Bitte prüfe deine Eingabe.",
                    "requestId": "0847b184-a073-488a-ad76-b0cd6f7d9554"
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "refreshToken": {
                    "type": "string"
                  }
                },
                "required": [
                  "refreshToken"
                ]
              }
            }
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "summary": "Sitzung und zugehörige Gerätetokens widerrufen",
        "operationId": "post_auth_logout",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "webCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Erfolgreich",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ok"
                }
              }
            }
          },
          "400": {
            "description": "Ungültige Eingabe",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Anmeldung erforderlich",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Keine Berechtigung",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Nicht gefunden oder gegenseitig gesperrt",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Versions- oder Idempotenzkonflikt",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Zu viele Anfragen",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Interner Fehler",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Fehler; keine privaten Daten.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": {
                    "code": "invalid_input",
                    "message": "Bitte prüfe deine Eingabe.",
                    "requestId": "0847b184-a073-488a-ad76-b0cd6f7d9554"
                  }
                }
              }
            }
          },
          "415": {
            "description": "Fehler; keine privaten Daten.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": {
                    "code": "invalid_input",
                    "message": "Bitte prüfe deine Eingabe.",
                    "requestId": "0847b184-a073-488a-ad76-b0cd6f7d9554"
                  }
                }
              }
            }
          },
          "503": {
            "description": "Fehler; keine privaten Daten.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": {
                    "code": "invalid_input",
                    "message": "Bitte prüfe deine Eingabe.",
                    "requestId": "0847b184-a073-488a-ad76-b0cd6f7d9554"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/reports": {
      "get": {
        "summary": "Meldungen; ausschließlich explizite Administratoren",
        "operationId": "get_admin_reports",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "webCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Erfolgreich",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Report"
                      }
                    }
                  },
                  "required": [
                    "items"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Ungültige Eingabe",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Anmeldung erforderlich",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Keine Berechtigung",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Nicht gefunden oder gegenseitig gesperrt",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Versions- oder Idempotenzkonflikt",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Zu viele Anfragen",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Interner Fehler",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Fehler; keine privaten Daten.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": {
                    "code": "invalid_input",
                    "message": "Bitte prüfe deine Eingabe.",
                    "requestId": "0847b184-a073-488a-ad76-b0cd6f7d9554"
                  }
                }
              }
            }
          },
          "415": {
            "description": "Fehler; keine privaten Daten.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": {
                    "code": "invalid_input",
                    "message": "Bitte prüfe deine Eingabe.",
                    "requestId": "0847b184-a073-488a-ad76-b0cd6f7d9554"
                  }
                }
              }
            }
          },
          "503": {
            "description": "Fehler; keine privaten Daten.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": {
                    "code": "invalid_input",
                    "message": "Bitte prüfe deine Eingabe.",
                    "requestId": "0847b184-a073-488a-ad76-b0cd6f7d9554"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      },
      "webCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-used-fruit-apple-auth"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message",
              "requestId"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "requestId": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "Tokens": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          },
          "expiresIn": {
            "type": "integer"
          }
        },
        "required": [
          "id",
          "accessToken",
          "refreshToken",
          "expiresIn"
        ],
        "description": "Opaque Zugangstokens, nicht lokal decodieren. Access 15 Minuten; Refresh 30 Tage. Jede Rotation ersetzt beide Tokens atomar in der Keychain. Wiederverwendung widerruft die Familie."
      },
      "Ok": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "enum": [
              true
            ]
          }
        },
        "required": [
          "ok"
        ]
      },
...
    }
  }
}